backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet==3.1.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.17
groq==0.13.0
httpx==0.28.1
email-validator==2.2.0
python-dotenv==1.0.1
pypdf==5.1.0
python-docx==1.1.2
